Split overrides out into smaller classes
